rename ops_*.py to operations_*.py
