+++ src/echodataflow/flows/flows_simulation.py
@@ -13,7 +13,7 @@ from botocore.config import Config
 from prefect import flow, runtime
 from prefect.variables import Variable
 
-from echodataflow.operations.ops_simulation import (
+from echodataflow.operations.operations_simulation import (
     S3CopyResult,
     S3CopySettings,
     S3CopyWorkItem,
