+++ tests/test_operations_copy_raw.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from echodataflow.operations.ops_simulation import (
+from echodataflow.operations.operations_simulation import (
     S3CopyResult,
     S3CopySettings,
     S3CopyWorkItem,
